Add supporting film option (#44)
* Ignore Sanctum migrations

* Add supportingFilmOf FK to screening migration

* Add select for supporting film

* Exclude supporting films in screening queries

* Remove unnecessary method in ScreeningController

* Add SupportingFilmNotes in lists

* Load supportingFilms for serial

* Rename SupportingFilmsNote

* Change handling of URL params

* Fix typo

* Update all URL params

* Add SupportingFilmsInfo to ScreeningPage

* Add MainFilmInfo

* Update SupportingFilmsInfo

* Update error handling in DeleteScreening method

* Update ScreeningsListItem

* Fix setting of doc title

* Fix condition in ScreeningsListItem

* Fix search result in ScreeningController

* Show titles of supporting films in overviews

* Add fields in ScreeningService

* Add date validation for supporting films

* Fix GetScreenings method

* Add updating of supporting film dates

* Refactor BillingController

* Fix name

* Run npm run prod

* Merge from main

* Update styles of overview page

* Fis mistakes found in review

+++ app/Http/Controllers/ScreeningController.php
@@ -6,10 +6,13 @@ use App\Http\Requests\ScreeningFormRequest;
 use App\Models\Screening;
 use App\Services\ImageService;
 use App\Services\ScreeningService;
+use DateTime;
+use Illuminate\Http\Exceptions\HttpResponseException;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Config;
 use Illuminate\Support\Facades\Storage;
+use Illuminate\Support\Facades\Validator;
 
 class ScreeningController extends Controller
 {
@@ -22,19 +25,25 @@ class ScreeningController extends Controller
         $this->imageService = $imageService;
     }
 
+    public function GetScreenings()
+    {
+        return Screening::whereNull('supportingFilmOf')->get();
+    }
+
     public function GetFutureScreenings()
     {
-        return Screening::where('date', '>', date("Y-m-d H:i:s"))->orderBy('date')->with('image')->get();
+        return Screening::where('date', '>', date("Y-m-d H:i:s"))
+            ->whereNull('supportingFilmOf')
+            ->orderBy('date')
+            ->with('image')
+            ->with('supportingFilms:supportingFilmOf,uuid,title')
+            ->get();
     }
 
     public function GetScreeningByUuid(string $uuid)
     {
-        return Screening::where('uuid', $uuid)->with('image.license')->with('serial')->first();
-    }
-
-    public function GetScreeningsByYear(int $year)
-    {
-        return Screening::whereYear('date', $year)->get();
+        return Screening::where('uuid', $uuid)
+            ->with('image.license')->with('serial')->with('mainFilm')->with('supportingFilms')->first();
     }
 
     public function GetScreeningsBySemester(string $semester)
@@ -55,11 +64,16 @@ class ScreeningController extends Controller
             ->orWhere('special', 'like', '%' . $search . '%')
             ->orWhere('author', 'like', '%' . $search . '%')
             ->orderByDesc('date')
+            ->with('supportingFilms:supportingFilmOf,title')
             ->get();
     }
 
     public function PostScreening(ScreeningFormRequest $request)
     {
+        if ($request->supportingFilmOf) {
+            $this->checkMainFilmDate($request);
+        }
+
         $screening = new Screening(['uuid' => uniqid(),]);
         $screening = $this->mapRequestToScreening($request, $screening);
 
@@ -73,7 +87,16 @@ class ScreeningController extends Controller
 
     public function PatchScreening(ScreeningFormRequest $request)
     {
-        $screening = Screening::where('uuid', $request->uuid)->first();
+        if ($request->supportingFilmOf) {
+            $this->checkMainFilmDate($request);
+        }
+
+        $screening = Screening::where('uuid', $request->uuid)->with('supportingFilms')->first();
+
+        if ($screening->supportingFilms) {
+            $this->updateSupportingFilmDates($screening, $request);
+        }
+
         $screening = $this->mapRequestToScreening($request, $screening);
         $screening->save();
         return $screening;
@@ -86,9 +109,14 @@ class ScreeningController extends Controller
         }
 
         $screening = Screening::firstWhere('uuid', $uuid);
+
         if ($screening->billing) {
             abort(422, 'Zu dieser Vorführung existiert noch eine Abrechnung, die vorher gelöscht werden muss.');
         }
+        if (count($screening->supportingFilms) > 0) {
+            abort(422, 'Diese Vorführung kann nicht gelöscht werden, weil ihr noch Vorfilme zugeordnet sind.');
+        }
+
         $image = $screening->image;
         $screening->delete();
         if ($image) {
@@ -118,8 +146,34 @@ class ScreeningController extends Controller
         $screening->special = $request->special;
         $screening->tercet = $request->tercet;
         $screening->serial_id = $request->serialId;
+        $screening->supportingFilmOf = $request->supportingFilmOf;
         $screening->author = $request->author;
 
         return $screening;
+    }
+
+    private function checkMainFilmDate(Request $request)
+    {
+        $mainFilm = Screening::where('id', $request->supportingFilmOf)->first();
+        $mainDate = new DateTime($mainFilm->date);
+        $validator = Validator::make($request->all(), [
+            'day' => 'date_equals:' . $mainDate->format('Y-m-d')
+        ]);
+        if ($validator->fails()) {
+            throw new HttpResponseException(
+                response()->json(
+                    ['validationErrors' => ['Ein Vorfilm muss das selbe Vorführdatum wie der Hauptfilm.']],
+                    422
+                )
+            );
+        }
+    }
+
+    private function updateSupportingFilmDates(Screening $screening, Request $request)
+    {
+        foreach ($screening->supportingFilms as $supportingFilm) {
+            $supportingFilm->date = $request->day . ' ' . $request->time;
+            $supportingFilm->save();
+        }
     }
 }
